Refactor to change pattern folder and inheritance tree.
